pyproject.toml
[project]
name = "maplestorage-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.11.14",
    "annotated-types==0.7.0",
    "asgiref==3.8.1",
    "celery==5.4.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.0",
    "django==5.1.4",
    "django-cors-headers==4.6.0",
    "django-redis>=5.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.5.0",
    "drf-yasg==1.21.10",
    "idna==3.10",
    "inflection==0.5.1",
    "mysqlclient>=2.2.7",
    "packaging==24.2",
    "playwright>=1.56.0",
    "psycopg2-binary>=2.9.10",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "pyjwt==2.9.0",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "python-dotenv==1.0.1",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "redis==5.2.1",
    "requests==2.32.3",
    "responses>=0.25.7",
    "sqlparse==0.5.3",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "uritemplate==4.1.1",
    "urllib3==2.2.3",
    "waitress>=3.0.2",
]
